Added neural gas, refactored things to be more general.
